refactor(swagger): move Swagger setup to a dedicated configuration file
