+++ src/main.ts
@@ -1,8 +1,7 @@
 import { ConfigService } from '@nestjs/config';
 import { NestFactory } from '@nestjs/core';
-import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
-import { apiReference } from '@scalar/nestjs-api-reference';
 import { AppModule } from './app.module';
+import { setupSwagger } from './config/docs';
 import type { Env } from './config/env';
 
 async function bootstrap() {
@@ -10,30 +9,22 @@ async function bootstrap() {
 
 	const configService = app.get<ConfigService<Env, true>>(ConfigService);
 
-	const port = configService.get('PORT', { infer: true });
+	/**
+	 * CORS
+	 */
 	const corsOrigin = configService.get('CORS_ORIGIN', { infer: true });
-
 	app.enableCors({ origin: corsOrigin ?? '*' });
 
 	/**
-	 *  SWAGGER
+	 * SWAGGER
+   -> documentação disponível em 'http://localhost:3333/api'
 	 */
-	const config = new DocumentBuilder()
-		.setTitle('API Projeto de Extensão ')
-		.setDescription('API do meu Projeto de extensão')
-		.setVersion('1.0.0')
-		.build();
+	setupSwagger(app);
 
-	const document = () => SwaggerModule.createDocument(app, config);
-	SwaggerModule.setup('api', app, document);
-
-	app.use(
-		'/reference',
-		apiReference({
-			content: document,
-		}),
-	);
-
+	/**
+	 * HTTP SERVER
+	 */
+	const port = configService.get('PORT', { infer: true });
 	await app.listen(port ?? 3333);
 }
 bootstrap();
